Fixed the cloud methods in tafparser.py
It properly parses the clouds from the TAFs and doesn't throw any
errors.

One thing I don't know about is in the CloudLayer class. I made it
possible to create a CloudLayer object with either a string from the TAF
or a pre-made CloudCoverage object. If you use a string as the
parameter, then it will use that string to make a new CloudCoverage.

I don't know if it's better to force the type one way or another, or to
allow the flexibility.