Consolidated venv checking to one place
The test for the presence of a Python virtual environment was duplicated
in runme.sh and run_tests.sh. This was bad application of the concept of Don't
Repeat Yourself. The test is now kept in activatevenv.sh which is
already sourced by runme.sh and run_tests.sh.

Additionally, with run_tests.sh, my Windows Python did not find modules
contained in the project directory. Adding $PWD to each member of PYTHONPATH
seemed to fix it.