Added test for wind headings in excess of 360
I adjusted the lark terminal for WIND_DIRECTION_DEGREES, which will
allow strings of 000 to 360, inclusive.

The test_winds_valid_heading function runs through headings of 0 to 999
and makes sure the correct ones get through while the incorrect ones do
not.

It seems wasteful to me... is it? I don't know.